@@ -2,26 +2,33 @@
   "name": "tutorbro-next",
   "private": true,
   "dependencies": {
-    "axios": "^0.25.0",
-    "express": "^4.15.4",
-    "google-libphonenumber": "^3.0.0",
-    "next": "^3.2.1",
+    "axios": "^1.6.8",
+    "express": "^4.19.2",
+    "google-libphonenumber": "^3.2.34",
+    "next": "^14.1.4",
     "nprogress": "^0.2.0",
-    "preact": "^8.2.5",
-    "preact-compat": "^3.17.0",
+    "preact": "^10.20.1",
+    "preact-compat": "^3.19.0",
     "prop-types": "^15.8.1",
-    "react": "^15.6.1",
-    "react-addons-css-transition-group": "^15.6.0",
-    "react-alert": "^2.2.0",
-    "react-dom": "^15.6.1",
-    "react-filepicker": "^0.10.16",
-    "react-ga": "^2.2.0",
-    "react-headroom": "^2.1.6",
+    "react": "^18.2.0",
+    "react-addons-css-transition-group": "^15.6.2",
+    "react-alert": "^7.0.3",
+    "react-dom": "^18.2.0",
+    "react-filepicker": "^0.11.0",
+    "react-ga": "^3.3.1",
+    "react-headroom": "^3.2.1",
     "react-helmet": "^6.1.0",
     "react-script-tag": "^1.1.2",
     "react-scrollspy": "^3.3.3",
     "react-spinjs": "^3.0.0",
-    "sw-precache-webpack-plugin": "^0.11.4"
+    "sw-precache-webpack-plugin": "^0.11.4",
+    "workbox-cacheable-response": "^7.0.0",
+    "workbox-core": "^7.0.0",
+    "workbox-expiration": "^7.0.0",
+    "workbox-precaching": "^7.0.0",
+    "workbox-routing": "^7.0.0",
+    "workbox-strategies": "^7.0.0",
+    "workbox-window": "^7.0.0"
   },
   "scripts": {
     "dev": "next",
@@ -33,8 +40,10 @@
     "analyze": "cross-env ANALYZE=1 next build"
   },
   "devDependencies": {
-    "cross-env": "^5.0.5",
-    "standard": "^10.0.3",
-    "webpack-bundle-analyzer": "^2.9.0"
+    "copy-webpack-plugin": "^12.0.2",
+    "cross-env": "^7.0.3",
+    "standard": "^17.1.0",
+    "webpack-bundle-analyzer": "^4.10.1",
+    "workbox-webpack-plugin": "^7.0.0"
   }
 }
