@@ -1,14 +1,18 @@
 // Require `PhoneNumberFormat`.
-var PNF = require('google-libphonenumber').PhoneNumberFormat
+import formate from "google-libphonenumber";
+
+var PNF = formate.PhoneNumberFormat;
 
 // Get an instance of `PhoneNumberUtil`.
-var phoneUtil = require('google-libphonenumber').PhoneNumberUtil.getInstance()
+import phoneUtil from "google-libphonenumber";
+
+const phoneUtils = phoneUtil.PhoneNumberUtil.getInstance();
 
 export default (mobilenumber, countryCode) => {
   // Parse number with country code.
-  var phoneNumber = phoneUtil.parse(mobilenumber, countryCode)
+  var phoneNumber = phoneUtils.parse(mobilenumber, countryCode);
 
   // return number in the international format.
-  return phoneUtil.format(phoneNumber, PNF.INTERNATIONAL)
+  return phoneUtils.format(phoneNumber, PNF.INTERNATIONAL);
   // => +1 202-456-1414
-}
+};
