@@ -1,18 +1,16 @@
-import Link from 'next/link'
-import Logo from './Logo'
-import NavLinks from './NavLinks'
-import ContactLinks from './ContactLinks'
+import Link from "next/link";
+import Logo from "./Logo";
+import NavLinks from "./NavLinks";
+import ContactLinks from "./ContactLinks";
 
-export default props => (
-  <nav className='nav'>
-    <div className='nav__brand'>
-      <Link href='/'>
-        <a>
-          <Logo />
-        </a>
+export default (props) => (
+  <nav className="nav">
+    <div className="nav__brand">
+      <Link href="/">
+        <Logo />
       </Link>
     </div>
-    <div className='nav__links'>
+    <div className="nav__links">
       <NavLinks />
       <ContactLinks {...props} />
     </div>
@@ -21,7 +19,7 @@ export default props => (
         .nav {
           display: flex;
         }
-        .nav__brandLogo{
+        .nav__brandLogo {
           width: 250px;
         }
         .nav__links {
@@ -33,4 +31,4 @@ export default props => (
       `}
     </style>
   </nav>
-)
+);
